+++ src/MusicDatabase1.java
@@ -1,11 +1,16 @@
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.Iterator;
+import java.util.ListIterator;
 import java.util.NoSuchElementException;
 
 /**
  * {@code MusicDatabase} represented as an {@link java.util.ArrayList
  * java.util.ArrayList} with implementations of primary methods.
+ *
+ * @convention No songs in this are null. There are no duplicate {@code Song}
+ *             objects in this. This only contains {@code Song} objects.
+ * @correspondence this is represented as an ArrayList of {@code Song} objects
  */
 public class MusicDatabase1 extends MusicDatabaseSecondary {
 
@@ -61,13 +66,13 @@ public class MusicDatabase1 extends MusicDatabaseSecondary {
     @Override
     public Iterator<Song> iterator() {
         Iterator<Song> it = new Iterator<Song>() {
-            private int pos = 0;
+            private int pos = -1;
+            private Boolean nextCalled = false;
 
             @Override
             public boolean hasNext() {
                 Boolean hasNextValue = false;
-                if (this.pos < MusicDatabase1.this.db.size()
-                        && MusicDatabase1.this.db.get(this.pos) != null) {
+                if ((this.pos + 1) < MusicDatabase1.this.db.size()) {
                     hasNextValue = true;
                 }
                 return hasNextValue;
@@ -79,9 +84,22 @@ public class MusicDatabase1 extends MusicDatabaseSecondary {
                     throw new NoSuchElementException(
                             "ERROR: No more values in this");
                 }
-                Song value = MusicDatabase1.this.db.get(this.pos);
                 this.pos++;
+                Song value = MusicDatabase1.this.db.get(this.pos);
+                this.nextCalled = true;
                 return value;
+            }
+
+            @Override
+            public void remove() {
+                assert this.nextCalled : ""
+                        + "Violation of: next() has been called at least "
+                        + "once before this method.";
+
+                MusicDatabase1.this.db.remove(this.pos);
+                this.pos--;
+
+                this.nextCalled = false;
             }
         };
         return it;
@@ -181,7 +199,6 @@ public class MusicDatabase1 extends MusicDatabaseSecondary {
                 }
             }
         }
-
         return list;
     }
 
@@ -189,9 +206,51 @@ public class MusicDatabase1 extends MusicDatabaseSecondary {
     @Override
     public ArrayList<Song> removeEntries(MusicDatabaseKernel.SearchField field,
             String value) {
-        // TODO Auto-generated method stub
-        throw new UnsupportedOperationException(
-                "Unimplemented method 'removeEntries'");
+        assert value != null : "Violation of: value != null";
+
+        ArrayList<Song> list = new ArrayList<Song>();
+
+        ListIterator<Song> dbIt = this.db.listIterator();
+
+        while (dbIt.hasNext()) {
+            Song song = dbIt.next();
+
+            switch (field) {
+                case TITLE: {
+                    if (song.title().equals(value)) {
+                        list.add(song);
+                        dbIt.remove();
+                    }
+                    break;
+                }
+                case ARTIST: {
+                    if (song.artist().equals(value)) {
+                        list.add(song);
+                        dbIt.remove();
+                    }
+                    break;
+                }
+                case ALBUM: {
+                    if (song.album().equals(value)) {
+                        list.add(song);
+                        dbIt.remove();
+                    }
+                    break;
+                }
+                case LENGTH: {
+                    if (song.length().equals(value)) {
+                        list.add(song);
+                        dbIt.remove();
+                    }
+                    break;
+                }
+                default: {
+                    break;
+                }
+            }
+        }
+
+        return list;
     }
 
     // CHECKSTYLE: ALLOW THIS METHOD TO BE OVERRIDDEN
@@ -236,7 +295,6 @@ public class MusicDatabase1 extends MusicDatabaseSecondary {
     // CHECKSTYLE: ALLOW THIS METHOD TO BE OVERRIDDEN
     @Override
     public void sort(Comparator<Song> order) {
-        // TODO Auto-generated method stub
-        throw new UnsupportedOperationException("Unimplemented method 'sort'");
+        this.db.sort(order);
     }
 }